logo-component: add logo component
everything that has to deal with logo, such as styles and images, should
be placed in a single Logo component

it also rename the files under the component folder to
- index.js
- style.css
as it is the pattern used in the Pilot project
