+++ src/components/Logo/images/banks/bradesco-black.png
