+++ src/components/Logo/images/banks/index.js
